test(parental-guide): support node test runner
